refactor: move src/ to adapters/vscode/
