+++ adapters/vscode/vscodeTerminalAdapter.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 
-import type { ITerminalAdapter, TerminalHandle } from '../core/src/terminalAdapter.js';
+import type { ITerminalAdapter, TerminalHandle } from '../../core/src/terminalAdapter.js';
 
 /** VS Code implementation of ITerminalAdapter. Wraps vscode.window terminal access. */
 export class VscodeTerminalAdapter implements ITerminalAdapter {
