+++ esbuild.js
@@ -76,7 +76,7 @@ const esbuildProblemMatcherPlugin = {
 
 async function main() {
   const ctx = await esbuild.context({
-    entryPoints: ['src/extension.ts'],
+    entryPoints: ['adapters/vscode/extension.ts'],
     bundle: true,
     format: 'cjs',
     minify: production,
